requirements.txt
asgiref==3.6.0
beautifulsoup4==4.12.2
boto3==1.26.135
botocore==1.29.135
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==3.1.0
cssselect2==0.7.0
distlib==0.3.6
Django==4.2.1
django-ckeditor==6.5.1
django-cleanup==7.0.0
django-crispy-forms==2.0
django-filer==2.2.4
django-js-asset==2.0.0
django-mptt==0.14.0
django-polymorphic==3.1.0
django-storages==1.13.2
easy-thumbnails==2.8.5
filelock==3.12.0
freetype-py==2.3.0
future==0.18.3
idna==3.4
jmespath==1.0.1
lxml==4.9.2
Pillow==9.5.0
platformdirs==3.5.0
pycairo==1.23.0
python-dateutil==2.8.2
qiniu==7.10.0
reportlab==4.0.0
requests==2.30.0
rlPyCairo==0.2.0
s3transfer==0.6.1
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
svglib==1.5.1
tinycss2==1.2.1
tzdata==2023.3
Unidecode==1.1.2
upload==0.1.1
urllib3==1.26.15
virtualenv==20.23.0
webencodings==0.5.1
